use-cases/customer-support/email-support-agent/requirements.txt
openai
commune-mail
